Updating the app to v. 1.0 of AiiDA
Due to the change in AiiDA core 1.0.0, there have been relevant changes
to the app. In particular

- The structure of the app is now much more modular; ideas where
shamelessly taken for hte AiiDA REST API v4 in order to reuse as much as
possible of the code

- Due to Flask architecture, it is not possible to trigger plumpy processes from
inside the app; all the relevant cases where it happens have been
wrapped into threaded decorated functions, that are actually executed in
a different thread, but waited to finish

- The app has now been `downgraded` to bare executioner of existing
workflows, and not in a preprocessing of structure followed by execution
of a workflow. This in the long run will be more flexible, and easier to
maintain without touching the code (more complex workflow can be of
course be nested in order to do exactly what I was doing before)

- The API is now truly an extension of the existing AiiDA rest: all of
the endpoints are accessible in order to use the excellent querying
ability already there

- The `resources` file contains several endpoints now;
  - the usual submit, that returns the uuid of the workflow upon
  successful execution, and additional info
  - The `properties` node wraps the logger and info of a process, to
  give the info in one location

TODO

- Some of the planned endpoints are still missing. Especially the one
looking for duplicate of the existing query (will probably end up being
a `?` option in submit as well) is important
- The endpoint looking for existing properties. This will depend on
another dictionary, so I know where to find the keywords (i.e. for a
`bandgap.pw` I know I should check for attribute `x` in an output_node
called `y`). This also to retrieve results
- Would be nice to provide, on failure to provide nice inputs, something
like the output of `verdi plugin list aiida.workflow xxx`, with the
required inputs, type, help. This should be relatively easy using
namespaces done like I fill the inputs
- Codes are now NOT accepted to be loaded from existing nodes, but needs
to be passed explicitely as inputs. It would be nice to not be obliged
to pass them, but use default ne. Unfortunately, if people do not add
the `Code` type specific to a calculation, this is impossible unless we
map all the workflow property with a schema to fill the blanks (we
should expect complex workflow to employ different code types in
different workflow points)
- General standardization of the code, that is a bit spaghetti in many
points, especially the error handling
- examples
- tests
-documentation
- inclusion of COD tools to filter the CIF files
- Having siesta and defect workflow up to date
-TESTS!