Fixes the workflow generation
Testing with much more complex workflows, I found issues that were not
working properly. Now this is fixed for the most complex workflow
example to date, that mixes all the possible (? I think) input types and
mixes in all the keywords.

Added:
- Better parsing of inputs
- Keyword to load a code from string (maybe useful for understanding
  what a code is)
- Removed ad-hoc validation of input (that is executed on workflow
  submission anyway, so it is still handled)
- Dynamic namespaces (when we do not know at the start how many inputs
  we have).

To fix:
- If dynamic namespaces can be ports (not sure they can or if it is
useful), that is not handled at the moment
- Tests still missing

To add:
- Endpoint for structure search
- Schema for output retrieval of specific property expected by the
  calculation
- Expand the documentation