simplified code, introduced properties-from-node search
Resource classes have been simplified by removing common parts into a
common base class (more can be done).

Introduced a new endpoint, existing/properties/prop/node, that returns
- any workflow that has been launched with <node> as input, and the
workflow related to <prop>
- a list of the properties actually in the database from node <node> and
related to the property <prop>

This endpdoint will be useful in order to know if it is necessary to
launch a calculation, if a usable value is there (the user can then
check the inputs used to generate it, to see if it is adequate) or if a
workflow is actualy running but not finished, or excepted.

TODO: still it would be nice to have a check in order to see if a
specific exact json submission dictionary alerady exist before launching

TODO: right now, given a workflow, I generate a "mock" workflow to
connect the json dictionary to the generated inputs of the workflow.
This means that I actually have a step inbetween: json -> mock ->
outputs -> workflow -> property. Now I cannot actually follow this path,
as json -> mock -> outputs -> can connect to *different* workflows. Due
to this, it might actually be better to introduce a workflow that CALLS
the other one, to be univoque. This, conversely, would work, but might
not be immediate to return the uuid of the workflow. Needs thinking and
testing