fixed structure.cod and added exit_statuses endpoint
The search.cod workflow had a problem if the get_aiida_structure method
was failing. This is now fixed, and what is returned is a list of aiida
structures, as well as a list of cif files that could not converted

Added the exit_statuses endpoint as /properties/<prop>/exit_statuses.
This endpoint returns the list of exit codes and the corresponding exit
messages returned by excepted, as requested by ginestra developers

TODO:

- update the documentation to reflect the changes
- re-discuss how to return a property, if by a dictionary with
name:property (eventually property to be not only a number, but a simple
python object) or directly in the API response in the form of a
dictionary